pyproject.toml
[project]
name = "canieatthemushroom"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "clipspy>=1.0.6",
    "pytest>=8.4.2",
    "reflex>=0.8.17",
    "google-genai>=1.47.0",
    "pillow>=10.0.0",
    "python-dotenv>=1.2.1",
]

[dependency-groups]
dev = [
    "marimo>=0.17.6",
]
